fix minor bugs (mostly display)

+++ inventorysystem/src/main/java/project/modules/user/Admin.java
@@ -4,6 +4,7 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Scanner;
 
+import project.global.ConsoleUI;
 import project.global.SqlConnector;
 import project.global.UserInputHandler;
 
@@ -26,96 +27,75 @@ public class Admin extends User {
 
     public void createAdmin()  // can work , validated  ,  redirected
     {
+        ConsoleUI.clearScreen();
         this.setUserId(super.generateUserId("A"));
 
-        System.out.println("Enter Admin Name: ");
-        this.setUserName(scanner.nextLine());
+        this.setUserName(UserInputHandler.getString("Enter Admin Name", 3));
 
-        System.out.println("Enter Admin Password: ");
-        this.setUserPassword(scanner.nextLine());
+        this.setUserPassword(UserInputHandler.getString("Enter Admin Password", "^(?=.*[@#$%^&+=])(?=\\S+$).{5,}$"));
 
-        while(!Validation.validateUserPassword(this.getUserPassword())) {
-            System.out.println("Format not allowed");
-            System.out.println("Enter Admin Password: ");
-            this.setUserPassword(scanner.nextLine());
-        }
-
-        System.out.println("Enter Admin Email: ");
+        
+        System.out.print("Enter Admin Email: ");
         this.setUserEmail(scanner.nextLine());
 
-        this.Add();
+        if(this.Add()){
 
-        System.out.println("Admin created sucessfully ");
+            System.out.println("Admin with userID " + this.getUserId() +" has been created sucessfully ");
 
-        System.out.println("Do you want to create another Admin? (Y/N): ");
-        String choice = scanner.nextLine();
-        
 
-        if (choice.equalsIgnoreCase("Y")) {
-            createAdmin();
+            String choice = UserInputHandler.getConfirmation("\nDo you want to create another Admin?");
 
-        } else {
-            UserMenu();
+            if (choice.equalsIgnoreCase("Y")) {
+                createAdmin();
+
+            }
+
+        }else {
+            System.out.println("Admin creation failed.");
         }
-
-
-
-       
+        
     }
     
    
 
     public void deleteAdmin() // can work , validated , X redirect
     {
-        do {
-
-            System.out.println("Enter Admin ID to delete: ");
-            this.setUserId(scanner.nextLine());
-            
-
-        } while (!Validation.validateUserId(this.getUserId()));
+        ConsoleUI.clearScreen();
+        displayAllUsers();
+        this.setUserId(UserInputHandler.getString("Enter Admin ID to delete", "^[ARV][0-9]{5}$"));
     
         if (!Get()) {
             System.out.println("Admin ID does not exist.");
-
-            System.out.println("Try again? (Y/N): ");
-            String choice = scanner.nextLine();
+            String choice = UserInputHandler.getConfirmation("Try again?");
 
             if (choice.equalsIgnoreCase("Y")) {
                 deleteAdmin();
-            } else {
-                redirectToMenu(scanner);
-            }
-
+            } 
             return;  
         } 
         else if (this.getUserId().equals(User.getLoggedInUserId())) {
             System.out.println("Error: You cannot delete your own account while logged in.");
-            redirectToMenu(scanner);
             return; 
         } 
     
-        displayUserDetails();
-    
-        System.out.println("Are you sure you want to delete this Admin (Y/N): ");
-        String choice = scanner.nextLine();
+        String choice = UserInputHandler.getConfirmation("\nAre you sure you want to delete this Admin");
     
         if (choice.equalsIgnoreCase("Y")) {
-            this.Remove();
-            System.out.println("Admin account deleted successfully.");
+            if(this.Remove()){
+                System.out.println("Admin account deleted successfully.");
 
-            System.out.println("Do you want to delete another Admin? (Y/N): ");
-            String choice2 = scanner.nextLine();
+                String choice2 = UserInputHandler.getConfirmation("\nDo you want to delete another Admin");
 
-            if (choice2.equalsIgnoreCase("Y")) {
-                deleteAdmin();
+                if (choice2.equalsIgnoreCase("Y")) {
+                    deleteAdmin();
+                }
+
             } else {
-                redirectToMenu(scanner);
+                System.out.println("Admin account deletion failed.");
             }
        } else {
             System.out.println("Operation cancelled.");
-            redirectToMenu(scanner);
-
+            ConsoleUI.pause();
         }
     
 
@@ -123,33 +103,23 @@ public class Admin extends User {
 
     public void UpdateAdmin() // can work
     {
-
-        do{
-                System.out.println("Enter Admin ID: ");
-                this.setUserId(scanner.nextLine());
-                
-        }while (!Validation.validateUserId(this.getUserId()));
+        ConsoleUI.clearScreen();
+        displayAllUsers();
+        this.setUserId(UserInputHandler.getString("\nEnter Admin ID", "^[ARV][0-9]{5}$"));
 
         if (!Get()){
 
             System.out.println("Admin ID does not exist.");
 
-            System.out.println("Try again? (Y/N): ");
-            String choice = scanner.nextLine();
+            String choice = UserInputHandler.getConfirmation("Try again?");
 
             if (choice.equalsIgnoreCase("Y")) {
                 UpdateAdmin();
-            } else {
-                redirectToMenu(scanner);
             }
 
         }
         
-
         boolean continueEditing = true;
-        
-        displayUserDetails();
-
 
         while (continueEditing) {
 
@@ -160,7 +130,6 @@ public class Admin extends User {
             System.out.println("3. Admin Email");
             int choice = UserInputHandler.getInteger("\nEnter choice: ", 1, 3);
             
-    
             String field;
             String value;
     
@@ -172,13 +141,8 @@ public class Admin extends User {
                     break;
                 case 2:
                     field = "Admin_Password";
-                    System.out.print("Enter new Admin Password: ");
-                    value = scanner.nextLine();
-    
-                    while (!Validation.validateUserPassword(value)) {
-                        System.out.println("Enter new Admin Password: ");
-                        value = scanner.nextLine();
-                    }
+                    value = UserInputHandler.getString("Enter new Admin Password", "^(?=.*[@#$%^&+=])(?=\\S+$).{5,}$");
+
                     break;
                 case 3:
                     field = "Admin_Email";
@@ -193,20 +157,17 @@ public class Admin extends User {
     
             super.Update(field, value);
     
-            System.out.println("Would you like to update another field? (Y/N): ");
-            String anotherFieldChoice = scanner.nextLine();
+            String anotherFieldChoice = UserInputHandler.getConfirmation("Would you like to update another field?");
 
             continueEditing = anotherFieldChoice.equalsIgnoreCase("Y");
         }
-
-        redirectToMenu(scanner);
     }
 
    
 
 
     @Override
-    public void Add()
+    public boolean Add()
     {
        String sql = "INSERT INTO Admin (Admin_Id, Admin_Name, Admin_Password, Admin_Email, Admin_Reg_Date) VALUES (?, ?, ?, ?, ?)";
 
@@ -215,10 +176,10 @@ public class Admin extends User {
         
         if(!Connector.isConnected()) {
             System.out.println("Connection failed");
-            return;
+            return false;
         }
 
-        Connector.PrepareExecuteDML(sql, this.getUserId(), this.getUserName(), this.getUserPassword(), this.getUserEmail(), LocalDateTime.now());
+        return Connector.PrepareExecuteDML(sql, this.getUserId(), this.getUserName(), this.getUserPassword(), this.getUserEmail(), LocalDateTime.now());
 
     }
 
@@ -229,47 +190,50 @@ public class Admin extends User {
 
         User user =  new Retailer();
         Retailer retailer  =  (Retailer) user ;
-  
-          System.out.println("1. Create Admin");
-          System.out.println("2. View Admin");
-          System.out.println("3. Update Admin");
-          System.out.println("4. Delete Admin");
-          System.out.println("5. Add Retailer");
-          System.out.println("6. Delete Retailer");
-          System.out.println("7. Notification Retailer");
-          System.out.println("8. Exit");
-          int choice = UserInputHandler.getInteger("\nEnter choice: ", 1, 8);
+        boolean exit = false;
+        while(!exit){
+            ConsoleUI.clearScreen();
+            System.out.println("Admin and Retailer Management");
+            System.out.println("1. Create Admin");
+            System.out.println("2. View Admin");
+            System.out.println("3. Update Admin");
+            System.out.println("4. Delete Admin");
+            System.out.println("5. Add Retailer");
+            System.out.println("6. Delete Retailer");
+            System.out.println("7. Notification Retailer");
+            System.out.println("8. Exit");
+            int choice = UserInputHandler.getInteger("\nEnter choice: ", 1, 8);
 
-  
-          switch (choice) {
-              case 1:
-                createAdmin();
-                break;
-              case 2:
-                displayAllUsers();
-                redirectToMenu(scanner);
-                break;
-              case 3:
-                UpdateAdmin();
-                break;
-              case 4: 
-                deleteAdmin();
-                  break;
-              case 5: 
-                retailer.createRetailer();      
-                  break;
-              case 6:
-                retailer.deleteRetailer();
-                break;
-            case 7:
-                Notification();
-                //implement send notification to retailer
-                break;
-            case 8:
-                return;
-            default:
-                System.out.println("Invalid choice.");
-                break;
+            switch (choice) {
+                case 1:
+                    createAdmin();
+                    break;
+                case 2:
+                    displayAllUsers();
+                    break;
+                case 3:
+                    UpdateAdmin();
+                    break;
+                case 4: 
+                    deleteAdmin();
+                    break;
+                case 5: 
+                    retailer.createRetailer();      
+                    break;
+                case 6:
+                    retailer.deleteRetailer();
+                    break;
+                case 7:
+                    Notification();
+                    //implement send notification to retailer
+                    break;
+                case 8:
+                    exit = true;
+                    break;
+                default:
+                    System.out.println("Invalid choice.");
+                    break;
+            }
         }
     }
 
@@ -277,7 +241,7 @@ public class Admin extends User {
         ArrayList<Request> pendingRequests = Request.viewRequest();
 
         if (pendingRequests != null && !pendingRequests.isEmpty()) {
-            int requestId = UserInputHandler.getInteger("Enter Request ID to further proceed: ", 1, Integer.MAX_VALUE);
+            int requestId = UserInputHandler.getInteger("\nEnter Request ID to further proceed: ", 0, Integer.MAX_VALUE);
 
             Request selectedRequest = null;
             for (Request req : pendingRequests) {
@@ -289,73 +253,34 @@ public class Admin extends User {
 
             // If the request is found, proceed with approval or rejection
             if (selectedRequest != null) {
-                System.out.println("Do you want to approve or reject the request (A/R): ");
-                String choice = scanner.next();
+                
+                String choice = UserInputHandler.getConfirmation("Do you want to approve or reject the request");
 
-                if (choice.equalsIgnoreCase("A")) {
+                if (choice.equalsIgnoreCase("Y")) {
                     selectedRequest.approveRequest(this.getUserId()); // Passing admin ID to approveRequest
-                    
-
-
-
                 } 
-                else if (choice.equalsIgnoreCase("R")) {
+                else if (choice.equalsIgnoreCase("N")) {
                     selectedRequest.rejectRequest();
-
-
                 } 
                 else {
                     System.out.println("Invalid choice.");
                 }
 
-                System.out.println("Do you want to handle more requests (Y/N): ");
-                String moreRequests = scanner.next();
+                String moreRequests = UserInputHandler.getConfirmation("Do you want to handle more requests");
 
                 if (moreRequests.equalsIgnoreCase("Y")) {
                     Notification(); // Recursively call the method to handle more requests
-                }
-                else if (moreRequests.equalsIgnoreCase("N")) {
-                    redirectToMenu(scanner);
-                }
-                else {
-                    System.out.println("Invalid choice.");
-                    redirectToMenu(scanner);
-                }
-
+                } 
             } else {
                 System.out.println("Request ID not found.");
-                redirectToMenu(scanner);
+                ConsoleUI.pause();
             }
 
         } else {
-            System.out.println("No pending requests found.");
-            redirectToMenu(scanner);
+            ConsoleUI.pause();
         }
-
-
         
     }
-
-
-    private void redirectToMenu(Scanner scanner) {
-        System.out.println("Do you want to return to the main menu? (Y/N): ");
-        String choice = scanner.nextLine();
-        if (choice.equalsIgnoreCase("Y")) {
-            UserMenu();
-        } else {
-            System.out.println("Exiting...");
-            System.exit(0);
-        }
-    }
-
-
-
-
-
-
-
-
-
 
 
 }
